refactor sanitization functions
