+++ sixel.go
@@ -42,7 +42,7 @@ func (sxs *sixelScreen) printSixel(win *win, screen tcell.Screen, reg *reg) {
 				break
 			}
 
-			line = sanitizeForDisplay(line)
+			line = sanitizePreview(line)
 			screen.LockRegion(win.x, y, printLength(line), 1, true)
 			fmt.Fprintf(&b, "\033[%d;%dH", y+1, win.x+1)
 			b.WriteString(line)
