+++ termseq.go
@@ -276,60 +276,3 @@ func applyOSC(body string, st tcell.Style) tcell.Style {
 		return st
 	}
 }
-
-// isControlChar reports whether a rune is a control character or otherwise
-// unsafe to display in a terminal.
-// Covers C0 (0x00-0x1F), DEL (0x7F), and C1 (0x80-0x9F).
-func isControlChar(r rune) bool {
-	return r < 0x20 || r == 0x7F || r >= 0x80 && r <= 0x9F
-}
-
-// sanitizeForDisplay replaces control characters and invalid bytes with the
-// Unicode replacement character (U+FFFD). Tabs are preserved for content
-// where tab expansion is handled by the renderer (e.g. preview panes).
-func sanitizeForDisplay(s string) string {
-	return strings.Map(func(r rune) rune {
-		if r == '\t' || !isControlChar(r) {
-			return r
-		}
-		return '\uFFFD'
-	}, s)
-}
-
-// sanitizeName sanitizes a filename, path, or symlink target for display.
-// Unlike sanitizeForDisplay it also replaces tabs, because tabs in names
-// are expanded by the renderer to tabstop width while displaywidth.String
-// counts them as width 1, causing column overflow.
-func sanitizeName(s string) string {
-	return strings.Map(func(r rune) rune {
-		if !isControlChar(r) {
-			return r
-		}
-		return '\uFFFD'
-	}, s)
-}
-
-// sanitizeMessage sanitizes a message intended for the message line. Like
-// sanitizeName it strips control runes, but it preserves terminal sequences
-// that lf itself recognizes (SGR, EL, OSC 8) so internal messages that use
-// color or hyperlinks still render correctly.
-func sanitizeMessage(s string) string {
-	s = strings.TrimRight(s, "\n\r")
-	var b strings.Builder
-	b.Grow(len(s))
-	for i := 0; i < len(s); {
-		if seq := readTermSequence(s[i:]); seq != "" {
-			b.WriteString(seq)
-			i += len(seq)
-			continue
-		}
-		r, w := utf8.DecodeRuneInString(s[i:])
-		if isControlChar(r) {
-			b.WriteRune('\uFFFD')
-		} else {
-			b.WriteRune(r)
-		}
-		i += w
-	}
-	return b.String()
-}
